Splittet opp All.js
Splittet opp All.js inn i forskjellige filer og konverterte det til typescript
