+++ oppgave_1/frontend/src/layout/Layout.tsx
@@ -1,0 +1,35 @@
+import { ReactNode } from 'react';
+
+interface LayoutProps {
+  children: ReactNode;
+}
+
+export default function Layout({ children }: LayoutProps) {
+  return (
+    <div
+      className="mx-auto grid min-h-screen w-full max-w-7xl grid-rows-[auto_minmax(900px,_1fr)_30px]"
+      data-testid="layout"
+    >
+      <nav className="mt-6 mb-12 flex justify-between">
+        <h1 className="text-lg font-bold uppercase" data-testid="logo">
+          <a href="/">Mikro LMS</a>
+        </h1>
+        <ul className="flex gap-8" data-testid="nav">
+          <li className="text-base font-semibold" data-testid="nav_courses">
+            <a href="kurs">Kurs</a>
+          </li>
+          <li className="text-base font-semibold" data-testid="nav_new">
+            <a href="/ny">Nytt kurs</a>
+          </li>
+        </ul>
+      </nav>
+      <main className="h-full">
+        {children}
+      </main>
+      <footer className="flex justify-between" data-testid="footer">
+        <p>Mikro LMS AS, 2024</p>
+        <p>99 00 00 00, mail@lms.no</p>
+      </footer>
+    </div>
+  );
+}
